Revert "Version world docs routes"
This reverts commit b52a351c3fb595042cb8c23dab97356aa11a2e67.
