--- docs/app/[lang]/v4/worlds/compare/page.tsx
@@ -1,10 +1,0 @@
-import {
-  compareWorldsMetadata,
-  VersionedCompareBenchmarksPage,
-} from '@/app/[lang]/_worlds/pages';
-
-export const metadata = compareWorldsMetadata;
-
-const Page = () => <VersionedCompareBenchmarksPage version="v4" />;
-
-export default Page;
