+++ docs/app/[lang]/v5/cookbook/[[...slug]]/page.tsx
@@ -26,7 +26,7 @@ import {
   rewriteCookbookUrlsInText,
 } from '@/lib/geistdocs/cookbook-source';
 import { getLLMText, getPageImage, v5Source } from '@/lib/geistdocs/source';
-import { buildVersionUrl, PRE_RELEASE_VERSION } from '@/lib/geistdocs/versions';
+import { PRE_RELEASE_VERSION } from '@/lib/geistdocs/versions';
 
 const VERSION_PREFIX = PRE_RELEASE_VERSION.prefix; // '/v5'
 
@@ -57,7 +57,7 @@ const Page = async ({
     let rewritten = rewriteCookbookUrlForVersion(href, VERSION_PREFIX);
     if (rewritten.startsWith('/docs/'))
       rewritten = `${VERSION_PREFIX}${rewritten}`;
-    return buildVersionUrl(rewritten, PRE_RELEASE_VERSION) as T;
+    return rewritten as T;
   }
   const RelativeLink = createRelativeLink(v5Source, publicPage);
   const V5CookbookLink = (props: ComponentProps<typeof RelativeLink>) => (
