--- docs/app/[lang]/v5/worlds/compare/page.tsx
@@ -1,10 +1,0 @@
-import {
-  compareWorldsMetadata,
-  VersionedCompareBenchmarksPage,
-} from '@/app/[lang]/_worlds/pages';
-
-export const metadata = compareWorldsMetadata;
-
-const Page = () => <VersionedCompareBenchmarksPage version="v5" />;
-
-export default Page;
