+++ docs/components/geistdocs/mobile-menu.tsx
@@ -3,9 +3,8 @@
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useEffect, useRef, useState } from 'react';
-import { IconArrowUpRight } from '@/components/geistcn-fallbacks/geistcn-assets/icons/icon-arrow-up-right';
 import { useVersion } from '@/hooks/geistdocs/use-version';
-import { buildVersionUrl } from '@/lib/geistdocs/versions';
+import { IconArrowUpRight } from '@/components/geistcn-fallbacks/geistcn-assets/icons/icon-arrow-up-right';
 import { cn } from '@/lib/utils';
 import { SearchButton } from './search';
 
@@ -85,11 +84,9 @@ export const MobileMenu = ({ items }: MobileMenuProps) => {
   const resolveHref = (href: string) => {
     if (
       !href.startsWith('http') &&
-      (href.startsWith('/docs') ||
-        href.startsWith('/cookbook') ||
-        href.startsWith('/worlds'))
+      (href.startsWith('/docs') || href.startsWith('/cookbook'))
     ) {
-      return buildVersionUrl(href, activeVersion);
+      return `${activeVersion.prefix}${href}`;
     }
     return href;
   };
