+++ docs/lib/geistdocs/versions.ts
@@ -25,15 +25,8 @@ export const VERSIONS: DocsVersion[] = [
   },
 ];
 
-const latestVersion = VERSIONS.find((v) => !v.preRelease);
-const preReleaseVersion = VERSIONS.find((v) => v.preRelease);
-
-if (!latestVersion || !preReleaseVersion) {
-  throw new Error('Docs versions must include latest and pre-release entries');
-}
-
-export const LATEST_VERSION = latestVersion;
-export const PRE_RELEASE_VERSION = preReleaseVersion;
+export const LATEST_VERSION = VERSIONS.find((v) => !v.preRelease)!;
+export const PRE_RELEASE_VERSION = VERSIONS.find((v) => v.preRelease)!;
 
 /**
  * Derive the active docs version from a pathname. Matches `/v5/...` (or
@@ -42,7 +35,7 @@ export const PRE_RELEASE_VERSION = preReleaseVersion;
  */
 export function getVersionFromPathname(pathname: string): DocsVersion {
   // The v5 segment sits either at the root (default locale hidden) or right
-  // after a locale segment - both cases are covered by checking positions
+  // after a locale segment — both cases are covered by checking positions
   // 0 and 1.
   const segments = pathname.split('/').filter(Boolean);
   if (segments[0] === 'v5' || segments[1] === 'v5') {
@@ -55,9 +48,9 @@ export function getVersionFromPathname(pathname: string): DocsVersion {
  * Build a URL for the same page under a different version. Preserves the
  * trailing path after `/docs/` and any locale prefix.
  *
- * `/docs`, `/cookbook`, and `/worlds` paths are version-specific. v4 docs
- * and cookbook routes keep their historical unprefixed URLs, while Worlds
- * pages use explicit `/v4/worlds` and `/v5/worlds` routes.
+ * Only `/docs/...` paths are version-specific. All other routes (e.g.
+ * `/cookbook`, `/worlds`) are shared across versions and are returned
+ * unchanged — there is no `/v5/cookbook` to navigate to.
  *
  * `usePathname()` can return either `/docs/...` (default locale hidden by
  * the i18n middleware) or `/<locale>/docs/...` (non-default locale shown).
@@ -68,36 +61,22 @@ export function buildVersionUrl(
   pathname: string,
   targetVersion: DocsVersion
 ): string {
-  if (!pathname.startsWith('/')) {
-    return pathname;
-  }
-
-  if (
-    !pathname.includes('/docs') &&
-    !pathname.includes('/cookbook') &&
-    !pathname.includes('/worlds')
-  ) {
+  // Worlds and other shared routes have no versioned equivalent — unchanged.
+  if (!pathname.includes('/docs') && !pathname.includes('/cookbook')) {
     return pathname;
   }
 
   const segments = pathname.split('/').filter(Boolean);
   // Structural segments are path tokens that are never locale prefixes.
   const isStructural = (s: string | undefined) =>
-    s === 'docs' ||
-    s === 'v4' ||
-    s === 'v5' ||
-    s === 'cookbook' ||
-    s === 'worlds';
+    s === 'docs' || s === 'v5' || s === 'cookbook';
   const localeSegments =
     segments[0] && !isStructural(segments[0]) ? segments.slice(0, 1) : [];
   let rest = segments.slice(localeSegments.length);
-  if (rest[0] === 'v4' || rest[0] === 'v5') rest = rest.slice(1);
-  const isWorldsRoute = rest[0] === 'worlds';
-  const prefixSegments = isWorldsRoute
-    ? [targetVersion.id]
-    : targetVersion.prefix
-      ? [targetVersion.prefix.replace(/^\//, '')]
-      : [];
+  if (rest[0] === 'v5') rest = rest.slice(1);
+  const prefixSegments = targetVersion.prefix
+    ? [targetVersion.prefix.replace(/^\//, '')]
+    : [];
   const joined = [...localeSegments, ...prefixSegments, ...rest].join('/');
   return `/${joined}`.replace(/\/+$/, '') || '/';
 }
