+++ docs/next.config.ts
@@ -100,22 +100,22 @@ const config: NextConfig = {
       // Redirect old world docs to new /worlds routes
       {
         source: '/docs/deploying/world/local-world',
-        destination: '/v4/worlds/local',
+        destination: '/worlds/local',
         permanent: true,
       },
       {
         source: '/docs/deploying/world/postgres-world',
-        destination: '/v4/worlds/postgres',
+        destination: '/worlds/postgres',
         permanent: true,
       },
       {
         source: '/docs/deploying/world/vercel-world',
-        destination: '/v4/worlds/vercel',
+        destination: '/worlds/vercel',
         permanent: true,
       },
       {
         source: '/docs/worlds',
-        destination: '/v4/worlds',
+        destination: '/worlds',
         permanent: true,
       },
       // Foundations "Common Patterns" page was retired in favor of dedicated
